pkg/query: eventually update rules client
Currently, a rule client for a store reference is only created
if a rules spec has been discovered at the time of the discovery of a store spec.

If a rules spec has been discovered after a store reference has been already created,
the rules spec is being ignored.

The effect is that we sometimes have discovered StoreAPI endpoints
without underlying RulesAPI endpoints.

This fixes it.

Signed-off-by: Sergiusz Urbaniak <sergiusz.urbaniak@gmail.com>